requirements.txt
numpy
tqdm
tiktoken
torch
datasets
huggingface-hub
hydra-core
wandb
rich
einops
streamlit